CPP Problem uploaded

+++ CPP/Program02.cpp
@@ -1,6 +1,6 @@
 ///////////////////////////////////////////////////////////
 //                                                       //
-//  File name :     Program01.cpp                        //
+//  File name :     Program02.cpp                        //
 //  Description :   Used to do Addition of two numbers   //
 //  Author :        Pallavi Omprakash Malewar            //
 //  Date :          05/09/2025                           //
@@ -11,16 +11,16 @@
 #include<iostream>
 using namespace std;
 
-int Addition(int No1, int No2)
+float Addition(float No1, float No2)
 {
-    int Ans;
+    float Ans;
     Ans = No1 + No2;
     return Ans;
 }
 
 int main()
 {
-    int Ret, Value1, Value2;
+    float Ret, Value1, Value2;
 
     cout<<"Enter first number : \n";
     cin>>Value1;
@@ -28,7 +28,7 @@ int main()
     cout<<"Enter second number : \n";
     cin>>Value2;
     
-    Ret = Addition(Value1,Value2);
+    Ret = Addition(Value1, Value2);
 
     cout<<"Addition is : "<<Ret<<"\n";
